emumを別ファイルに定義
